Add a script for generating docs and pushing them to gh-pages
Simply run ./generate_docs, this will run `lein codox`, create a new commit on
the `gh-pages` branch, and push it to `origin`.

The script goes out of it way not to mess with your repo. It uses a separate
directory as a temporary working tree, and creates the commit without switching
branches.